Create Temp sketch
Adapted W.A. Smith's code to read the voltage from a TMP36 temperature sensor on pin A0 of the Arduino. Converts the voltage to a temperature and displays it on an LCD (16 character by two line).  Added code to convert to Fahrenheit. Added code to display a temp setpoint on the LCD. Added code to adjust setpoint with buttons. Added code to compare setpoint value temp and  turn on LED if below set point value. This simulates activating an SSR to control a heating element which would control temperature.